GC Lao files from CAT, Google Docs (#125)
Follow up to #88. I've added the machine translated files in a
`001_machineraw` folder. That might not be a good idea, but I couldn't
think of what else to call it.

I wasn't sure how we wanted to define the comments for segmentation.
I've added segmentation comments but I'm not sure if this gets us what
we need or not.

There were a few places where the heading was not at the beginning in
the paragraph. In these cases my script outputs things as they were in
the xliff. We can clean those up as we have time or when editing.

This PR also adds a basic xliff-to-markdown parser that I used initially
to get the data from the cat into markdown files.