Remove files removed on main (#178)
This is to deal with the merge conflict on #89. I think this
should resolve the merge conflict, since it replicates the change on
main. Once this is merged to latex-pdfs I think that branch should be
ready to review and merge to main. After that we can send PRs to main
and work on the remaining improvements as normal.