Generate PDFs with LaTeX (#89)
This adds basic support for generating PDFs with LaTeX running in GitHub
Actions. `draftall.sh` is still the main entry point, and basically the
only changes to the old scripts are related to directory paths and
markdown extension (`.text` -> `.md`). With this PR artifacts are
generated on every commit, but we could narrow that to only commits
touching `.tooling/latex` or `*.md` files. 

This resolves the following issues:
- [x] Make the paths dynamic instead of hardcoded to PP1/PP01 #94
- [x] Remove the reference code headings in the output PDF #95 
- [x] Fix poetry #174

---------

Co-authored-by: Tony Quick <59699423+tonyquicktech@users.noreply.github.com>
Co-authored-by: Brian Wilson <15809934+bountonw@users.noreply.github.com>