ICML generation for LBF (#239)
This PR adds support for automatically generating ICML (Adobe InCopy) files for the LBF book. ICML files can be imported into InDesign to allow externally-controlled content in InDesign files. The scripts in this PR convert the Markdown source files to ICML using Pandoc and a number of intermediate transformations with Lua scripts. Some of these scripts are LBF-specific and will need to be extracted when the process is expanded to support other books.

---------

Co-authored-by: Brian Wilson <15809934+bountonw@users.noreply.github.com>