Move tooling into external repo (#247)
This uses the new external tooling repository
(https://github.com/bountonw/translate-tooling/pull/1) for linting, PDF
generation, and ICML generation. The main workflow impact of this PR is
that changes to the forbidden words lists will have to be pushed from
the git submodule into the translate-tooling repo. I can help talk
through how that will change things when we're ready to merge this.