Update author info (#291)
The spelling of Waggoner has been confirmed by Kosin. There is a Thai author with the last name of Waggoner that spells it this way. (Of note, there is an online Jehovah's Witness article in Thai seeking to debunk certain Adventist understands which references Waggoner and spells it แวกกอเนอร์, which was the spelling I had originally suggested several years ago when Mike asked me how to spell it. But on further review, I think Kosin's suggestion is better.)