Proofed through chapter 21
The blue boxes and texts are marking issues that need to be reported or
changes that need to be made in the source text. Green is for things
that I thought might be issues but that Duang said were ok. Red is for
issues that have been reported in the current report.