LBF edits during final typesetting (#360)
- Fixes #357 
- Fixes #358 

This round of typesetting turned out to be an edit, with over 40 chapter
files changed. Changes were made in the source files here, and in the
.indb (InDesign) files that are not currently being tracked. All edits
have been marked in [the attached changes
PDF.](th/LBF/04_pdf/LBF_202407221300_changes-brian.pdf) For some reason,
GitHub is showing it as an invalid PDF, so I hope that it can be opened
and that it shows the changes. The annotations were made in Mac Preview.
The blue colored boxes noted paragraphs that I wanted to look at in
InDesign. These have been addressed, either by tweaking the typesetting
to my satisfaction or in deciding that what we have is the best we are
going to get.

**Note:** @kenton-r page 112 is now blank due to editing that freed up
space for the extra line.

---------

Co-authored-by: Tony Quick <59699423+tonyquicktech@users.noreply.github.com>