Add indd file for Thai MB (#366)
This file was originally named MB v 5.80 (World Paragraph Composer).indd
and was finalized sometime in January 2023. However, we've opted to use
a Christmas 2022 timestamp for this historical file. This version was
not the one sent to the printers for the first edition; instead, it
served as a working copy for generating sharable PDFs. It will also be
the basis for a new typeset to align with updated source files and
address issues in the current PDFs and print books.