Add main Lao dictionary (#556)
This is an initial attempt at a Lao word list to be used in typesetting and spell checking etc. All terms are in one word list. In the future, we may decide to split the terms into separate lists.

The first column is the clean term. the `|` is used to indicate the end of the term. The second column has weighted break points inserted for LaTeX pdf rendering. The `%` is used to make comments. I have used a very conservative weighting for these breakpoints. The file will be adjusted when we begin using it in typesetting. 

Terms are included from GC01-GC04, and GC16, along with some random terms gleaned while editing other chapters. 

---------

Co-authored-by: Brian Wilson <15809934+bountonw@users.noreply.github.com>