Initial Lao md to pdf typesetting pipeline (#575)
The pipeline is working. There are a few extra scripts that need to be
moved or deleted. There is also a ton of refactoring that needs to be
done and some bug squashing. But, as things are basically working. I am
moving it to main.

Calling `bash scripts/make_pdf.sh GC##` from the `04_assets` folder runs all 3 prep modules in debug mode, outputs several log files and produces the pdf of the chapter called.

The magic is in the dictionaries with up to six levels of dictionary overrides. So far only main, patch, book, and chapter are being used. But book_patch and chapter_patch are available.

Two tex files control the latex header and still need a lot of adjustment. But that will happen during the typesetting process.


---------

Co-authored-by: Matthew Leffler <mattleff@gmail.com>