Fix typos while processing GC02-GC16 (#648)
Merging to main for the Sabbath. GC17 has been started. This is a first
pass through the book, with the goal of working through all `\nodict{}`
tags by fixing typos and adding words to various dictionaries. It is
surprising how many typos have been caught. I also did some
troubleshooting in order to get certain files to compile and to allow
for multiple component terms to be tagged properly.

---------

Co-authored-by: Matthew Leffler <mattleff@gmail.com>