Handle log_folder in module2
Writes logs to folder based on time to 04_assets/pdf/logs.

Appends to files there instead of overwriting/clearing files so that we can keep it between runs of make_book.sh.

No longer overwrites lookahead_decisions but does write timestamp into file so you know when it did things. lookahead_decisions is not put in the log folder properly, but writing to that file is buried so deep I did not see it as worth the time since it does not have a lot written to it at this time.

Various refactoring to do things without overall breaking scripts.