TOC now obeys commands (#672)
Had to shift tex_header_info to be processed after the TOC is shown. Something in there is messing with the spacing on the TOC. To shift that down, had to shift some logic to make the TOC file (.toc) not have section information in it whatsoever. So a few things got shifted around in the Python scripts and in the docclass file.

Might be a problem if we need to have more \lw{} etc. use on a title page or similar before the TOC. Maybe we can avoid that somehow?

Also adjusts head height and font size to fix some errors showing up in the console.